+++ R/quadmesh-package.R
@@ -1,0 +1,15 @@
+#' @keywords internal
+"_PACKAGE"
+
+#' World raster map
+#'
+#' A rasterized version of `wrld_simpl`.
+#' @docType data
+#' @name worldll
+NULL
+#' World topography map
+#'
+#' A simplified version of 'Etopo2'.
+#' @docType data
+#' @name etopo
+NULL